Search PCAP with GPU/requirements.txt
scapy>=2.5.0
numpy>=1.24.0
cupy-cuda12x>=12.0.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
psutil>=5.9.0
colorama>=0.4.6
